fit span relocation
previously when it encounters a grapheme cluster
it should find the floor/ceil position of the
continuation byte to index into the span string

this uses a copy of two unstable functions from std:
`str::floor_char_boundary`
`str::ceil_char_boundary`
which are tracked here: https://github.com/rust-lang/rust/issues/93743

when this becomes stable, we can just rely on std.
but for now they are fine. (even if they are
obviously inefficient)