Add sorting and grouping.
This introduces 2 new flags:
-s, --sort <file|lint|default>
-g, --group <file|lint|default>

These flags are exclusive.

When using --sort the output will be sorted by the kind.
When using --group the output will be sorted and the output will be adjusted.

This also adds a --mode flag.
--mode <record|report|default>

With this flag, you can record (via --mode record) the raw output of the command, saved locally in the current working directory as .ccs.json. Then you can "replay" that session via --mode replay. This allows you to tweak colors, see renderer outputs or submit bug reports.

Breaking:
This changes the configuration slightly. It adds a new color key: 'position' which is the `:fragment` in `file:col:row`. Be sure to add a color for this or you'll get a runtime error about a missing value from the configuration file.